requirements.txt
asgiref==3.8.1
Django==5.0.6
Markdown==3.6
sqlparse==0.5.0
tzdata==2024.1
python-dotenv==1.0.1
pillow==10.4.0
python-telegram-bot==21.4
djangorestframework==3.15.2
isort==6.0.1
black==25.1.0
flake8==7.1.2
